prepare expo-router, remove some unnecessary files
