+++ starters/next-expo-router/apps/expo/app/_layout.tsx
@@ -1,0 +1,37 @@
+import { DarkTheme, DefaultTheme, ThemeProvider } from '@react-navigation/native'
+import { Provider } from 'app/provider'
+import { useFonts } from 'expo-font'
+import { Stack, useRootNavigation } from 'expo-router'
+import { useColorScheme } from 'react-native'
+
+export default function HomeLayout() {
+  const [loaded] = useFonts({
+    Inter: require('@tamagui/font-inter/otf/Inter-Medium.otf'),
+    InterBold: require('@tamagui/font-inter/otf/Inter-Bold.otf'),
+  })
+  const scheme = useColorScheme()
+
+  if (!loaded) {
+    return null
+  }
+  return (
+    <Provider>
+      <ThemeProvider value={scheme === 'dark' ? DarkTheme : DefaultTheme}>
+        <Stack>
+          <Stack.Screen
+            name="index"
+            options={{
+              title: 'Home',
+            }}
+          />
+          <Stack.Screen
+            name="user/[id]"
+            options={{
+              title: 'User',
+            }}
+          />
+        </Stack>
+      </ThemeProvider>
+    </Provider>
+  )
+}
