+++ starters/next-expo-router/apps/expo/app/user/[id].tsx
@@ -1,0 +1,3 @@
+import { UserDetailScreen } from 'app/features/user/detail-screen'
+
+export default UserDetailScreen
